requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
dictdiffer==0.9.0
fastapi==0.112.2
frozenlist==1.4.1
h11==0.14.0
idna==3.8
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.5
psycopg2==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
starlette==0.38.2
tomlkit==0.13.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
uvicorn==0.30.6
yarl==1.9.6
